Move BaseNetwork to it's own pattern
